# Convert pool query to function call