## Changes proposed in this PR
- Convert pool query logic to pure function call

## Why are we making these changes?
- Easier to read
- Easier to reuse
- Better performance
```
New (no pool)
Evaluation count : 83280 in 60 samples of 1388 calls.
             Execution time mean : 724.906725 µs
    Execution time std-deviation : 7.126763 µs
   Execution time lower quantile : 717.701292 µs ( 2.5%)
   Execution time upper quantile : 740.208994 µs (97.5%)
                   Overhead used : 1.565043 ns

Found 3 outliers in 60 samples (5.0000 %)
	low-severe	 1 (1.6667 %)
	low-mild	 2 (3.3333 %)
 Variance from outliers : 1.6389 % Variance is slightly inflated by outliers


Old (no pool)
Evaluation count : 40980 in 60 samples of 683 calls.
             Execution time mean : 1.470291 ms
    Execution time std-deviation : 8.615948 µs
   Execution time lower quantile : 1.453152 ms ( 2.5%)
   Execution time upper quantile : 1.486338 ms (97.5%)
                   Overhead used : 1.565043 ns


New (has pool)
Evaluation count : 54660 in 60 samples of 911 calls.
             Execution time mean : 1.090170 ms
    Execution time std-deviation : 16.439733 µs
   Execution time lower quantile : 1.067333 ms ( 2.5%)
   Execution time upper quantile : 1.124302 ms (97.5%)
                   Overhead used : 1.565043 ns

Found 5 outliers in 60 samples (8.3333 %)
	low-severe	 4 (6.6667 %)
	low-mild	 1 (1.6667 %)
 Variance from outliers : 1.6389 % Variance is slightly inflated by outliers

Old (has pool)
Evaluation count : 35460 in 60 samples of 591 calls.
             Execution time mean : 1.754258 ms
    Execution time std-deviation : 51.372349 µs
   Execution time lower quantile : 1.706617 ms ( 2.5%)
   Execution time upper quantile : 1.871717 ms (97.5%)
                   Overhead used : 1.565043 ns

Found 9 outliers in 60 samples (15.0000 %)
	low-severe	 6 (10.0000 %)
	low-mild	 3 (5.0000 %)
 Variance from outliers : 15.8231 % Variance is moderately inflated by outliers
```